requirements.txt
numpy
opencv-python
face_recognition
scipy
scikit-learn
fastapi
uvicorn
python-multipart
sqlalchemy
requests
tensorflow-cpu
